Use LoginAuth() to exchange a username/password for auth token
After a new OAuthSession type is created with NewOAuthSession(),
the caller is given the choice of how the auth token should be
created. The auth token is required to define the HTTP client used
for API requests.

Currently, LoginAuth() is the only method of token exchange. Other
methods should be implemented in the future, which is why LoginAuth()
does not happen by default within the constructor of OAuthSession.